+++ server/package.json
@@ -1,6 +1,6 @@
 {
   "name": "cityvizor-server",
-  "version": "2.0.0",
+  "version": "4.2.0",
   "description": "",
   "homepage": "https://cityvizor.cz/",
   "repository": "otevrena-data-mfcr/CityVizor",
